+++ sourcefiles/csv.h
@@ -2,6 +2,6 @@
 #define CSV_SEPARATOR		','
 
 int csvSize(char* filename, int& M, int& N);
-int csvRead(char* filename, long double** mat, int M, int N);
-int csvWrite(char* filename, long double** mat, int M, int N);
+int csvRead(char* filename, double** mat, int M, int N);
+int csvWrite(char* filename, double** mat, int M, int N);
