+++ sourcefiles/getFileInfo.h
@@ -5,5 +5,5 @@
 enum typeOfData {DATE=0, VALUE};
 
 int fileSampleSize(char* filename, int& M, int& N);
-int fileSampleRead(char* filename, long double** mat, int M, int N);
-int fileSampleWrite(char* filename, long double** mat, int M, int N);
+int fileSampleRead(char* filename, double** mat, int M, int N);
+int fileSampleWrite(char* filename, double** mat, int M, int N);
